refactor(sender) : move external function to utils
